Model setup, some tooling, environment management, etc. The basic contexts and models have been added: Accounts, Events, and Inventory.
Accounts is for user and organization management, Events is for creating and managing events, and Inventory is for lists of things that are global (in this case, Games, Categories, and Series).

Also added some tooling for database migrations akin to Rake's `db:migrate` and such. Should be nicer than having to invoke `crystal run src/db_migrate.cr -- ...` every time. Now it's just `./bin/tool db:migrate`.

Also now using Dotenv to load environment variables instead of having them set on my computer. Only took 5 years to get here, but we've made it.